back/requirements.txt
fastapi>=0.127.0
uvicorn[standard]>=0.40.0

# ORM (Pydantic + SQLAlchemy)
sqlmodel>=0.0.31

# Postgres driver (SQLAlchemy "postgresql+psycopg://")
psycopg[binary]>=3.3.2

# Settings
pydantic-settings>=2.12.0

# Auth
bcrypt>=5.0.0
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.21

# Redis for pub/sub
redis>=5.2.0

# Payment processing
stripe>=14.1.0

# Image processing
Pillow>=12.1.0

# HTTP requests for wine import
requests>=2.32.5

# Email sending
aiosmtplib>=3.0.0

# PDF generation
reportlab>=4.4.7

# Excel export for reports
openpyxl>=3.1.0
